refactor: rename cruiser manager to state service
